+++ src/main.rs
@@ -1,52 +1,14 @@
+mod components;
 mod constants;
 mod systems;
 
+use crate::components::*;
 use crate::constants::*;
 use crate::systems::{
     collision::collision_system, input::keyboard_input_system, round::round_system,
     velocity::velocity_system,
 };
 use bevy::{prelude::*, render::pass::ClearColor};
-
-pub struct Game {
-    left_score: usize,
-    right_score: usize,
-    prev_winner: Option<Player>,
-}
-impl Default for Game {
-    fn default() -> Self {
-        Self {
-            left_score: 0,
-            right_score: 0,
-            prev_winner: None,
-        }
-    }
-}
-// Player Type
-#[derive(Debug)]
-pub enum PlayerType {
-    Left,
-    Right,
-}
-#[derive(Debug)]
-pub struct Player {
-    player_type: PlayerType,
-}
-
-#[derive(Debug)]
-pub struct Velocity(Vec2);
-pub struct Ball;
-pub struct LostRound;
-
-pub enum Collidable {
-    Reflect, // Something that is collidable but reflects the ball
-    End,     // Something that is collidable but ends the balls movement
-}
-#[derive(Debug)]
-pub enum WallSide {
-    Left,
-    Right,
-}
 
 fn main() {
     App::build()
